core: Add Uid
Add the Uid interface to be represent all the current *Uid classes.

This will be used in a subsequent commit to make RetroRunner take a Uid
parameter rather than a UID string, allowing to evaluate the UIDs
lazily and hence helping to construct a RetroRunner earlier with no
performance impact.

This is needed to create a generic game type.