retro-runner: Turn '*_path' props into getter methods
This allows the evaluation of the paths to throw errors, which is needed
to allow lazy computation of the UID used in the paths as it can throw
errors, and hence helping to construct a RetroRunner earlier with no
performance impact.

This is needed to create a generic game type.