Fix module path checking
The code first checks for the retro core the legacy way, for
compatibility.

If it doesn't find anything, then it's supposed to search for the core
by mimetypes.

However, the Retro.search_module funciton returns null when the module
is not found, and creating a GFile with a null path raises an assertion.

In addition, we can use the FileUtils API which is nicer for us, as it
removes the need to create a GFile.

https://bugzilla.gnome.org/show_bug.cgi?id=769909