retro: Don't try to load the state if its size is wrong
The core is supposed to return serialized data of a certain, expected
size.

Unfortunately, they often don't.

When that happens, we probably shouldn't try to load the serialized
state, as it is potentially bad.

Unfortunately, it might also be correct.

If we don't load such serialized data and it was correct, the player
will lose the state of their game, which is pretty bad.

However if we try loading such serialized data and it was bad, then the
core will just completely freeze the whole application, with no way for
the user to even close it.

The second option seems even worse, so with this commit we throw an
error when the size wasn't the expected one.

https://bugzilla.gnome.org/show_bug.cgi?id=769938