ui: Improve the resume dialog
"Main menu" is very unclear in this context: is that the main menu of
the Games application or the main menu of the game?

"Title screen" might be a more descriptive wording, but at the same time
it is usually better to use verbs for buttons.

"Restart" seems like a decent verb to use here.

https://bugzilla.gnome.org/show_bug.cgi?id=769924