ui: Drop the cancel button when starting a game
The player just asked us to start a game, why would they want to cancel?

This seems like such a rare use-case that we don't need a button here,
especially as there's a workaround: they could just resume, and click
the back arrow to get back to the home page.

This makes the resume dialog less heavy, as it felt very crowded with 3
buttons for just a short line of text.

https://bugzilla.gnome.org/show_bug.cgi?id=769924