ui: Inhibit logout while playing
This allows notifying the user when they try to logout that a game was
running, giving them a chance to cancel the logout, go back to the game,
save their progress, play for an hour or two because they got caught
back in, save their progress again and then finally logout.

This also handles poweroff, since it first logs out.

https://bugzilla.gnome.org/show_bug.cgi?id=769968