retro: Be more tolerant about serialization sizes
We are having problems with Nestopia, which returns 21400 the first
time, and then 21380 every time after that.

With the current code, which checks for equality, we always fail to
deserialize the Nestopia saved states.

The libretro.h header has this to say about the function:

    // Returns the amount of data the implementation requires to serialize internal state (save states).
    // Beetween calls to retro_load_game() and retro_unload_game(), the returned size is never allowed to be larger than a previous returned value, to
    // ensure that the frontend can allocate a save state buffer once.

They seem to implicitly acknowledge that this value can change, but that
it can only ever be smaller than a previous value.

This commit changes the check for equality into a check for inferiority,
which makes loading a saved state for Nestopia work.

https://bugzilla.gnome.org/show_bug.cgi?id=770783