ui: Use MIME types and factories froom plugins
Use the MIME types and URI game factories provided by plugins in the
application to search for game resources and to build games.

This will be used in the next commits to allow plugins to stop providing
game sources, helping to split the conerns of looking for game resources
and building games from them.

https://bugzilla.gnome.org/show_bug.cgi?id=781334