utils: Fix options injection with custom filename
If a filename start by '--', grep would assume this is a option,
and so would potentially either miss the file, or wait
until STDIN is closed. A quick look at grep manpage do not
yield any way to execute code with this for now, but it is
better to fix the problem for now.

Using '--' is the traditional way to deal with file using
a double dash prefix.

Fix https://bugzilla.gnome.org/show_bug.cgi?id=782952